+++ fluentnamer/src/main/java/com/azure/autorest/fluent/transformer/FluentTransformer.java
@@ -41,7 +41,7 @@ public class FluentTransformer {
     }
 
     public CodeModel postTransform(CodeModel codeModel) {
-        codeModel = new LongRunningOperationProcessor(fluentJavaSettings.getJavaNamesForRemoveOperationGroup()).process(codeModel);
+        codeModel = new OperationGroupFilter(fluentJavaSettings.getJavaNamesForRemoveOperationGroup()).process(codeModel);
         codeModel = new NamingConflictResolver().process(codeModel);
         codeModel = new SchemaRenamer(fluentJavaSettings.getJavaNamesForRenameModel()).process(codeModel);
         codeModel = new OperationNameNormalization().process(codeModel);
